third_party/blink/web_tests/http/tests/xmlhttprequest/range-test-expected.txt

Test Range support in XMLHttpRequest

getRange(resources/reply.xml, 34, 36, false)
Length : expected 2 got 2
getRange(resources/reply.xml, 35, 38, false)
Length : expected 3 got 3
getRange(resources/reply.xml, 34, 36, true)
Length : expected 2 got 2
getRange(resources/reply.xml, 35, 38, true)
Length : expected 3 got 3
